refactor: changed color parser from `tinycolor2` to `color` (Qix-)
This is a breaking change since the color API will change substancially,
but will fix #203, an issue that was a problem for a long time in v2
and it also was in v3, because of tinycolor that was losing color
information on conversion (like the hue).

Qix Color is a very good parser with no color info loss.